api/requirements.txt
fastapi
uvicorn
selenium
webdriver-manager
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
